@@ -1,46 +1,35 @@
 package com.javadevjournal.customer.data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+import lombok.Data;
 
+//import javax.persistence.*;
+@Data
 @Entity
 public class CustomerModel {
 
     @Id
-    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
     private Integer id;
-    @Column(unique = true,nullable = false)
-    private String email;
-    private String firstName;
-    private String lastName;
+//    @Column(unique = false,nullable = false)
+    private int user_id;
+    private int item_id;
+    private int rating;
+    private int time_stamp;
 
-    protected CustomerModel() {}
+    public CustomerModel() {}
 
-    public CustomerModel(String email,String firstName, String lastName) {
-        this.email=email;
-        this.firstName = firstName;
-        this.lastName = lastName;
+    public CustomerModel(int user_id, int item_id, int rating, int time_stamp) {
+        this.user_id = user_id;
+        this.item_id = item_id;
+        this.rating = rating;
+        this.time_stamp = time_stamp;
     }
 
     @Override
     public String toString() {
         return String.format(
-                "Customer[id=%d, firstName='%s', lastName='%s',email='%s']",
-                id, firstName, lastName,email);
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
+                "Movie[id=%d, user_id='%s', item_id='%s',rating='%s', time_stamp='%s']",
+                id, user_id, item_id, rating, time_stamp);
     }
 }
