@@ -1,48 +1,40 @@
 package com.javadevjournal.product.data;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+//import javax.persistence.*;
+
+@Data
 @Entity
+@NoArgsConstructor
 public class ProductModel {
 
     @Id
-    @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private int id;
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
+    private Integer id;
+//    @Column(unique = false,nullable = false)
+    private int user_id;
+    private int item_id;
+    private int rating;
+    private int time_stamp;
 
-    @Column(nullable = false)
-    private String code;
-    private String name;
-    private double price;
 
 
-    protected ProductModel() {}
+//    protected ProductModel() {}
 
-    public ProductModel(String code, String name, double price) {
-        this.code = code;
-        this.name = name;
-        this.price = price;
+    public ProductModel(int user_id, int item_id, int rating, int time_stamp) {
+        this.user_id = user_id;
+        this.item_id = item_id;
+        this.rating = rating;
+        this.time_stamp = time_stamp;
     }
 
     @Override
     public String toString() {
         return String.format(
-                "Product[id=%d, code='%s', name='%s', price='%s']",
-                id, code, name,price);
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public double getPrice() {
-        return price;
+                "Movie[id=%d, user_id='%s', item_id='%s',rating='%s', time_stamp='%s']",
+                id, user_id, item_id, rating, time_stamp);
     }
 }
