@@ -1,6 +1,8 @@
 package com.javadevjournal.customer.config;
 
 
+import jakarta.persistence.EntityManagerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.jdbc.DataSourceBuilder;
@@ -14,45 +16,44 @@ import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-import javax.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
 @Configuration
 @EnableTransactionManagement
 @EnableJpaRepositories(
-        entityManagerFactoryRef = "customerEntityManagerFactory",
-        transactionManagerRef = "customerTransactionManager",
+        entityManagerFactoryRef = "oracleEntityManagerFactory",
+        transactionManagerRef = "oracleTransactionManager",
         basePackages = {"com.javadevjournal.customer.repo"}
 )
 public class CustomerConfig {
 
     @Primary
-    @Bean(name = "customerDataSource")
+    @Bean(name = "oracleDataSource")
     @ConfigurationProperties(prefix = "spring.datasource")
-    public DataSource customerDataSource() {
+    public DataSource oracleDataSource() {
         return DataSourceBuilder.create().build();
     }
 
     @Primary
-    @Bean(name = "customerEntityManagerFactory")
+    @Bean(name = "oracleEntityManagerFactory")
     public LocalContainerEntityManagerFactoryBean
     entityManagerFactory(
             EntityManagerFactoryBuilder builder,
-            @Qualifier("customerDataSource") DataSource dataSource
+            @Qualifier("oracleDataSource") DataSource dataSource
     ) {
         return builder
                 .dataSource(dataSource)
                 .packages("com.javadevjournal.customer.data")
-                .persistenceUnit("db1")
+//                .persistenceUnit("db1")
                 .build();
     }
 
     @Primary
-    @Bean(name = "customerTransactionManager")
-    public PlatformTransactionManager customerTransactionManager(
-            @Qualifier("customerEntityManagerFactory") EntityManagerFactory
-                    customerEntityManagerFactory
+    @Bean(name = "oracleTransactionManager")
+    public PlatformTransactionManager oracleTransactionManager(
+            @Qualifier("oracleEntityManagerFactory") EntityManagerFactory
+                    oracleEntityManagerFactory
     ) {
-        return new JpaTransactionManager(customerEntityManagerFactory);
+        return new JpaTransactionManager(oracleEntityManagerFactory);
     }
 }
